Add a version of the proxy that can be run locally
